toolkit/content/tests/chrome/test_tooltip_noautohide.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>

<window title="Tooltip Noautohide Tests"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>      
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>      

<tooltip id="thetooltip" noautohide="true"
         onpopupshown="setTimeout(tooltipStillShown, 6000)"
         onpopuphidden="ok(gChecked, 'tooltip did not hide'); SimpleTest.finish()">
  <label id="label" value="This is a tooltip"/>
</tooltip>

<button id="button" label="Tooltip Text" tooltip="thetooltip"/>

<script class="testbody" type="application/javascript">
<![CDATA[

var gChecked = false;

function runTests()
{
  var button = document.getElementById("button");
  var windowUtils = window.windowUtils;
  windowUtils.disableNonTestMouseEvents(true);
  synthesizeMouse(button, 2, 2, { type: "mouseover" });
  synthesizeMouse(button, 4, 4, { type: "mousemove" });
  synthesizeMouse(button, 6, 6, { type: "mousemove" });
  windowUtils.disableNonTestMouseEvents(false);
}

function tooltipStillShown()
{
  gChecked = true;
  document.getElementById("thetooltip").hidePopup();
}

SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(runTests);

]]>
</script>

<body xmlns="http://www.w3.org/1999/xhtml">
<p id="display">
</p>
<div id="content" style="display: none">
</div>
<pre id="test">
</pre>
</body>

</window>
